# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API tasks",
        "description": "Management tasks",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/"
        }
    ],
    "paths": {
...
        "/api/tasks": {
            "get": {
                "tags": [
                    "tasks"
                ],
                "summary": "List Tasks by user",
                "operationId": "5f4c89efc1b4ff38441eb44aa9b4e023",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "response": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "messages": {
                                            "type": "list",
                                            "example": "[...]"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": 1
                                                    },
                                                    "id_user": {
                                                        "type": "number",
                                                        "example": 1
                                                    },
                                                    "id_status": {
                                                        "type": "number",
                                                        "example": 1
                                                    },
                                                    "id_priority": {
                                                        "type": "number",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "text"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "text"
                                                    },
                                                    "date_limit": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2023-02-23T12:33:45.000000Z"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2023-02-23T12:33:45.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2023-02-23T12:33:45.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "response": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "messages": {
                                            "type": "list",
                                            "example": "[...]"
                                        },
                                        "data": {
                                            "type": "list",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/tasks/create": {
            "post": {
                "tags": [
                    "tasks"
                ],
                "summary": "Tasks create",
                "operationId": "1822fba8b80d191dc7d595eb47b4de52",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "number",
                                        "example": 1
                                    },
                                    "priority": {
                                        "type": "number",
                                        "example": 1
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "New Task"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Task description"
                                    },
                                    "date_limit": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-10-15T14:30:00.000000Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "response": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "messages": {
                                            "type": "list",
                                            "example": "[...]"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": 1
                                                    },
                                                    "id_user": {
                                                        "type": "number",
                                                        "example": 1
                                                    },
                                                    "id_status": {
                                                        "type": "number",
                                                        "example": 1
                                                    },
                                                    "id_priority": {
                                                        "type": "number",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "text"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "text"
                                                    },
                                                    "date_limit": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2023-02-23T12:33:45.000000Z"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2023-02-23T12:33:45.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2023-02-23T12:33:45.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "response": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "messages": {
                                            "type": "list",
                                            "example": "[...]"
                                        },
                                        "data": {
                                            "type": "list",
                                            "example": "[]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}